make URL=http://localhost:8080/davbox/jdo check  -> running `basic':  0. init.................. pass  1. begin................. pass  2. options............... pass  3. put_get............... WARNING: length mismatch: 0 vs 41     ...................... pass (with 1 warning)  4. put_get_utf8_segment.. WARNING: length mismatch: 0 vs 41     ...................... pass (with 1 warning)  5. put_no_parent......... pass  6. mkcol_over_plain...... FAIL (MKCOL on plain resource `/davbox/jdo/litmus/res-%e2%82%ac' succeeded!)  7. delete................ pass  8. delete_null........... pass  9. delete_fragment....... pass 10. mkcol................. pass 11. mkcol_again........... pass 12. delete_coll........... pass 13. mkcol_no_parent....... pass 14. mkcol_with_body....... pass 15. finish................ pass <- summary for `basic': of 16 tests run: 15 passed, 1 failed. 93.8% -> 2 warnings were issued. See debug.log for network/debug traces. make: *** [check] Error 1

+++ src/main/java/cc/co/llabor/dav/jdo/Jdo4Dav.java
@@ -68,6 +68,7 @@ public class Jdo4Dav extends AbstractTransactionalDaver implements IWebdavStore 
 			// delete file
 			try{
 				MemoryFileItem toDel = memFS.get(uri); 
+				if (toDel==null)return;
 				System_out_println("<delete>\n\t<file name=\'"+toDel.getName()+"\'/>... ");
 				toDel.delete();
 				Object retval = memFS.delete(toDel);
@@ -142,14 +143,10 @@ public class Jdo4Dav extends AbstractTransactionalDaver implements IWebdavStore 
  		String[] list = memFS.list(folderUri);
  		String[] listNoDot = new String[list.length-1];
  		int i=0;
- 		try{
-	 		for (String val:list){
-	 			if (!".".equals( val )){
-	 				listNoDot [i++]=val;
-	 			}
-	 		}
- 		}catch(java.lang.ArrayIndexOutOfBoundsException e){
- 			listNoDot = new String[]{"."};
+ 		for (String val:list){
+ 			if (!".".equals( val )){
+ 				listNoDot [i++]=val;
+ 			}
  		}
 		return  listNoDot;
 	}
